Switch to using Ints for answers
Ints are by default 64 bits on Apple platforms, so using a Double doesn't actually provide more fidelity.

+++ Sources/DailySolver.swift
@@ -10,6 +10,6 @@ import Foundation
 public protocol DailySolver {
     associatedtype CalculationInput
     func ParseInput(_ input: String) -> CalculationInput
-    func PerformPart1Calculation(_ input: CalculationInput) -> Double?
-    func PerformPart2Calculation(_ input: CalculationInput) -> Double?
+    func PerformPart1Calculation(_ input: CalculationInput) -> Int?
+    func PerformPart2Calculation(_ input: CalculationInput) -> Int?
 }
